@@ -3,62 +3,62 @@ require "fileutils"
 
 module Octopolo
   describe Config do
-    let(:cli) { mock("cli") }
+    let(:cli) { double("cli") }
 
     context "#initialize" do
       subject { Config }
 
       it "loads plugins" do
-        subject.any_instance.should_receive(:load_plugins)
+        expect_any_instance_of(subject).to receive(:load_plugins)
         subject.new(deploy_branch: "foo", branches_to_keep: ["a", "b"])
       end
 
       it "sets up methods for all the attributes it receives" do
         config = subject.new(deploy_branch: "foo", branches_to_keep: ["a", "b"])
 
-        config.deploy_branch.should == "foo"
-        config.branches_to_keep.should == ["a", "b"]
+        expect(config.deploy_branch).to eq("foo")
+        expect(config.branches_to_keep).to eq(["a", "b"])
       end
     end
 
     context "default cuzomizable methods" do
       context "#deploy_branch" do
         it "is master by default" do
-          Config.new.deploy_branch.should == "master"
+          expect(Config.new.deploy_branch).to eq("master")
         end
 
         it "is the specified branch otherwise" do
-          Config.new(deploy_branch: "production").deploy_branch.should == "production"
+          expect(Config.new(deploy_branch: "production").deploy_branch).to eq("production")
         end
       end
 
       context "#branches_to_keep" do
         it "is an empty array by default" do
-          Config.new.branches_to_keep.should == []
+          expect(Config.new.branches_to_keep).to eq([])
         end
 
         it "is the specified values otherwise" do
-          Config.new(branches_to_keep: ["a", "b"]).branches_to_keep.should == ["a", "b"]
+          expect(Config.new(branches_to_keep: ["a", "b"]).branches_to_keep).to eq(["a", "b"])
         end
       end
 
       context "#deploy_methods" do
         it "is an empty array by default" do
-          Config.new.deploy_methods.should == []
+          expect(Config.new.deploy_methods).to eq([])
         end
 
         it "is the specified values otherwise" do
-          Config.new(deploy_methods: ["a", "b"]).deploy_methods.should == ["a", "b"]
+          expect(Config.new(deploy_methods: ["a", "b"]).deploy_methods).to eq(["a", "b"])
         end
       end
 
       context "#deploy_environments" do
         it "is an empty array by default" do
-          Config.new.deploy_environments.should == []
+          expect(Config.new.deploy_environments).to eq([])
         end
 
         it "is the specified values otherwise" do
-          Config.new(deploy_environments: ["a", "b"]).deploy_environments.should == ["a", "b"]
+          expect(Config.new(deploy_environments: ["a", "b"]).deploy_environments).to eq(["a", "b"])
         end
       end
 
@@ -78,23 +78,23 @@ module Octopolo
         end
 
         it "returns the specified value otherwise" do
-          Config.new(github_repo: "tstmedia/ngin").github_repo.should == "tstmedia/ngin"
+          expect(Config.new(github_repo: "tstmedia/ngin").github_repo).to eq("tstmedia/ngin")
         end
       end
 
       context "#merge_resolver" do
         it "is nil by default" do
-          Config.new.merge_resolver.should == nil
+          expect(Config.new.merge_resolver).to eq(nil)
         end
 
         it "returns a string if it has a value" do
-          Config.new(merge_resolver: "/opt/resolver.sh").merge_resolver.should == "/opt/resolver.sh"
+          expect(Config.new(merge_resolver: "/opt/resolver.sh").merge_resolver).to eq("/opt/resolver.sh")
         end
       end
 
       context "#user_notifications" do
         it "is nil by default" do
-          Config.new.user_notifications.should == nil
+          expect(Config.new.user_notifications).to eq(nil)
         end
 
         it "raise an error if it is not an array" do
@@ -102,19 +102,19 @@ module Octopolo
         end
 
         it "returns the specified value if an array" do
-          Config.new(user_notifications: ["NickLaMuro"]).user_notifications.should == ["NickLaMuro"]
+          expect(Config.new(user_notifications: ["NickLaMuro"]).user_notifications).to eq(["NickLaMuro"])
         end
 
         it "returns the specified value if as an array if a string" do
-          Config.new(user_notifications: "NickLaMuro").user_notifications.should == ["NickLaMuro"]
+          expect(Config.new(user_notifications: "NickLaMuro").user_notifications).to eq(["NickLaMuro"])
         end
       end
 
       context "#plugins" do
-        before { Config.any_instance.stub(:load_plugins) }
+        before { allow_any_instance_of(Config).to receive(:load_plugins) }
 
         it "defaults to an empty array" do
-          Config.new.plugins.should == []
+          expect(Config.new.plugins).to eq([])
         end
 
         it "raise an error if it is not a string or array" do
@@ -122,37 +122,27 @@ module Octopolo
         end
 
         it "returns the specified single plugin as an array" do
-          Config.new(plugins: "octopolo-templates").plugins.should == ["octopolo-templates"]
+          expect(Config.new(plugins: "octopolo-templates").plugins).to eq(["octopolo-templates"])
         end
 
         it "returns the specified plugins as an array" do
-          Config.new(plugins: ["op-templates", "op-pivotal"]).plugins.should == ["op-templates", "op-pivotal"]
-        end
-      end
-
-      context "#use_pivotal_tracker" do
-        it "defaults to false" do
-          expect(Config.new.use_pivotal_tracker).to be_false
-        end
-
-        it "forces a truthy value to be true" do
-          expect(Config.new(use_pivotal_tracker: "true").use_pivotal_tracker).to be_true
+          expect(Config.new(plugins: ["op-templates"]).plugins).to eq(["op-templates"])
         end
       end
 
       context "#use_jira" do
         it "defaults to false" do
-          expect(Config.new.use_jira).to be_false
+          expect(Config.new.use_jira).to be_falsey
         end
 
         it "forces a truthy value to be true" do
-          expect(Config.new(use_jira: "true").use_jira).to be_true
+          expect(Config.new(use_jira: "true").use_jira).to be_truthy
         end
       end
 
       context "#jira_user" do
         it "does not raise an exception if jira isn't enabled" do
-          expect { Config.new.jira_user }.to_not raise_error(Config::MissingRequiredAttribute)
+          expect { Config.new.jira_user }.to_not raise_error
         end
 
         it "raises an exception if not given" do
@@ -166,7 +156,7 @@ module Octopolo
 
       context "#jira_password" do
         it "does not raise an exception if jira isn't enabled" do
-          expect { Config.new.jira_password }.to_not raise_error(Config::MissingRequiredAttribute)
+          expect { Config.new.jira_password }.to_not raise_error
         end
 
         it "raises an exception if not given" do
@@ -179,7 +169,7 @@ module Octopolo
       end
       context "#jira_url" do
         it "does not raise an exception if jira isn't enabled" do
-          expect { Config.new.jira_url }.to_not raise_error(Config::MissingRequiredAttribute)
+          expect { Config.new.jira_url }.to_not raise_error
         end
 
         it "raises an exception if not given" do
@@ -202,7 +192,7 @@ module Octopolo
           end
 
           it "includes any monkey patching" do
-            subject.example_var.should == ExamplePlugin::EXAMPLE_CONSTANT
+            expect(subject.example_var).to eq(ExamplePlugin::EXAMPLE_CONSTANT)
           end
         }
       end
@@ -213,7 +203,7 @@ module Octopolo
         end
 
         it "not include any monkey patching" do
-          expect{ subject.example_var.should }.to raise_error NoMethodError
+          expect{ expect(subject.example_var).to }.to raise_error NoMethodError
         end
       end
     end
@@ -223,8 +213,8 @@ module Octopolo
       subject { Config }
 
       it "reads from the .octopolo.yml file and creates a new config instance" do
-        subject.should_receive(:attributes_from_file).and_return(parsed_attributes)
-        subject.should_receive(:new).with(parsed_attributes)
+        expect(subject).to receive(:attributes_from_file).and_return(parsed_attributes)
+        expect(subject).to receive(:new).with(parsed_attributes)
 
         subject.parse
       end
@@ -235,8 +225,8 @@ module Octopolo
       subject { Config }
 
       it "parses the YAML in the octopolo_config_path" do
-        subject.stub(:octopolo_config_path).and_return(stub_path)
-        subject.attributes_from_file.should == YAML.load_file(stub_path)
+        allow(subject).to receive(:octopolo_config_path).and_return(stub_path)
+        expect(subject.attributes_from_file).to eq(YAML.load_file(stub_path))
       end
     end
 
@@ -246,8 +236,8 @@ module Octopolo
       before { project_working_dir }
 
       it "gives up if it can't find a config file" do
-        File.stub(:exists?) { false }
-        Octopolo::CLI.should_receive(:say).with("*** WARNING: Could not find .octopolo.yml or .automation.yml ***")
+        allow(File).to receive(:exists?) { false }
+        expect(Octopolo::CLI).to receive(:say).with("*** WARNING: Could not find .octopolo.yml or .automation.yml ***")
         subject.octopolo_config_path
         Dir.chdir project_working_dir
       end
@@ -259,13 +249,13 @@ module Octopolo
 
         it "is the .octopolo.yml file in the project directory" do
           Dir.chdir "spec/support"
-          subject.octopolo_config_path.should == File.join(Dir.pwd, '.octopolo.yml')
+          expect(subject.octopolo_config_path).to eq(File.join(Dir.pwd, '.octopolo.yml'))
         end
 
         it "is the .octopolo.yml file in the project directory, two directories up" do
           FileUtils.mkdir_p "spec/support/tmp/foo"
           Dir.chdir "spec/support/tmp/foo"
-          subject.octopolo_config_path.should == File.join(Dir.pwd, '.octopolo.yml')
+          expect(subject.octopolo_config_path).to eq(File.join(Dir.pwd, '.octopolo.yml'))
         end
 
         after do
@@ -282,13 +272,13 @@ module Octopolo
 
         it "is the .octopolo.yml file in the project directory" do
           Dir.chdir "spec/support"
-          subject.octopolo_config_path.should == File.join(Dir.pwd, '.automation.yml')
+          expect(subject.octopolo_config_path).to eq(File.join(Dir.pwd, '.automation.yml'))
         end
 
         it "is the .octopolo.yml file in the project directory, two directories up" do
           FileUtils.mkdir_p "spec/support/tmp/foo"
           Dir.chdir "spec/support/tmp/foo"
-          subject.octopolo_config_path.should == File.join(Dir.pwd, '.automation.yml')
+          expect(subject.octopolo_config_path).to eq(File.join(Dir.pwd, '.automation.yml'))
         end
 
         after do
@@ -314,7 +304,7 @@ module Octopolo
 
         it "skips loading the plugin and displays a message" do
           subject.instance_variable_set(:@plugins, "not-a-real-plugin")
-          subject.should_receive(:puts)
+          expect(subject).to receive(:puts)
                  .with("Plugin 'not-a-real-plugin' failed to load")
 
           subject.load_plugins
@@ -324,7 +314,7 @@ module Octopolo
 
     context "#remote_branch_exists?" do
       before do
-        Octopolo::CLI.stub(:perform => <<-BR
+        allow(Octopolo::CLI).to receive_messages(:perform => <<-BR
                           * origin/production
                             origin/test
                             origin/asdf
@@ -333,11 +323,11 @@ module Octopolo
       end
 
       it "should find production" do
-        subject.remote_branch_exists?("production").should == true
+        expect(subject.remote_branch_exists?("production")).to eq(true)
       end
 
       it "shouldn't production" do
-        subject.remote_branch_exists?("not-there").should == false
+        expect(subject.remote_branch_exists?("not-there")).to eq(false)
       end
     end
 
@@ -352,7 +342,7 @@ module Octopolo
         Dir.chdir "spec/support"
         config = Config.new
         expected_value = File.basename(File.dirname(Config.octopolo_config_path))
-        config.basedir.should == expected_value
+        expect(config.basedir).to eq(expected_value)
       end
 
       after do
