@@ -12,68 +12,68 @@ module Octopolo
 
   describe CLIWrapper do
     let(:foo) { Foo.new }
-    let(:cli) { stub(:CLI) }
+    let(:cli) { double(:CLI) }
 
     context "#cli" do
       it "wraps around the CLI class" do
-        foo.cli.should == CLI
+        expect(foo.cli).to eq(CLI)
       end
 
       it "uses the given CLI class" do
         foo.cli = cli
-        foo.cli.should == cli
+        expect(foo.cli).to eq(cli)
       end
     end
   end
 
   describe ConfigWrapper do
     let(:foo) { Foo.new }
-    let(:config) { stub(:config) }
+    let(:config) { double(:config) }
 
     context "#config" do
       it "parses the current config" do
-        Octopolo.should_receive(:config) { config }
-        foo.config.should == config
+        expect(Octopolo).to receive(:config) { config }
+        expect(foo.config).to eq(config)
       end
 
       it "uses the given parsed config" do
-        Config.should_not_receive(:parse)
+        expect(Config).not_to receive(:parse)
         foo.config = config
-        foo.config.should == config
+        expect(foo.config).to eq(config)
       end
     end
   end
 
   describe UserConfigWrapper do
     let(:foo) { Foo.new }
-    let(:user_config) { stub(:user_config) }
+    let(:user_config) { double(:user_config) }
 
     context "#user_config" do
       it "parses the current user config" do
-        UserConfig.should_receive(:parse) { user_config }
-        foo.user_config.should == user_config
+        expect(UserConfig).to receive(:parse) { user_config }
+        expect(foo.user_config).to eq(user_config)
       end
 
       it "uses the given parsed config" do
         foo.user_config = user_config
-        UserConfig.should_not_receive(:parse)
-        foo.user_config.should == user_config
+        expect(UserConfig).not_to receive(:parse)
+        expect(foo.user_config).to eq(user_config)
       end
     end
   end
 
   describe GitWrapper do
     subject { Foo.new }
-    let(:git) { stub(:Git) }
+    let(:git) { double(:Git) }
 
     context "#git" do
       it "wraps around the Git class" do
-        subject.git.should == Git
+        expect(subject.git).to eq(Git)
       end
 
       it "uses the given Git" do
         subject.git = git
-        subject.git.should == git
+        expect(subject.git).to eq(git)
       end
     end
   end
