@@ -7,54 +7,54 @@ module Octopolo
       context ".new login" do
         it "remembers the given login" do
           user = User.new "foo"
-          user.login.should == "foo"
+          expect(user.login).to eq("foo")
         end
       end
 
       context "#author_name" do
-        let(:octo) { stub }
+        let(:octo) { double }
         let(:login) { "joeuser" }
         let(:user) { User.new login }
 
         before do
-          user.stub(user_data: octo)
+          allow(user).to receive_messages(user_data: octo)
         end
 
         it "fetches the real name from GitHub" do
-          octo.stub(name: "Joe User")
-          user.author_name.should == octo.name
+          allow(octo).to receive_messages(name: "Joe User")
+          expect(user.author_name).to eq(octo.name)
         end
 
         it "returns the login if GitHub user has no name" do
-          octo.stub(name: nil)
-          user.author_name.should == user.login
+          allow(octo).to receive_messages(name: nil)
+          expect(user.author_name).to eq(user.login)
         end
       end
 
       context "#user_data" do
         let(:login) { "joeuser" }
         let(:user) { User.new login }
-        let(:octo) { stub }
+        let(:octo) { double }
 
         it "fetches the data from the User class" do
-          User.should_receive(:user_data).with(login) { octo }
-          user.user_data.should == octo
+          expect(User).to receive(:user_data).with(login) { octo }
+          expect(user.user_data).to eq(octo)
         end
       end
 
       context ".user_data login" do
         let(:base_login) { "joeuser" }
-        let(:octo) { stub }
+        let(:octo) { double }
 
         it "fetches the data from GitHub" do
           login = "#{base_login}#{rand(100000)}"
-          GitHub.should_receive(:user).with(login) { octo }
-          User.user_data(login).should == octo
+          expect(GitHub).to receive(:user).with(login) { octo }
+          expect(User.user_data(login)).to eq(octo)
         end
 
         it "caches the data" do
           login = "#{base_login}#{rand(100000)}"
-          GitHub.should_receive(:user).once { octo }
+          expect(GitHub).to receive(:user).once { octo }
           User.user_data(login)
           User.user_data(login)
         end
