@@ -9,26 +9,26 @@ module Octopolo
       let(:template_contents) { "<%= foo %>" }
 
       it "renders the given template name with the given local variables" do
-        Renderer.should_receive(:contents_of).with(template_name) { template_contents }
+        expect(Renderer).to receive(:contents_of).with(template_name) { template_contents }
         rendered = Renderer.render(template_name, locals)
-        rendered.should == locals[:foo]
+        expect(rendered).to eq(locals[:foo])
       end
     end
 
     context ".template_base_path" do
       it "should be in the octopolo directory" do
-        Renderer.template_base_path.should == File.expand_path(File.join(__FILE__, "../../../lib/octopolo/templates"))
+        expect(Renderer.template_base_path).to eq(File.expand_path(File.join(__FILE__, "../../../lib/octopolo/templates")))
       end
     end
 
     context ".contents_of template" do
-      let(:contents) { stub(:erb) }
+      let(:contents) { double(:erb) }
       let(:name) { "some_template" }
       let(:path) { File.join(Renderer.template_base_path, "#{name}.erb") }
 
       it "reads the file from templates/" do
-        File.should_receive(:read).with(path) { contents }
-        Renderer.contents_of(name).should == contents
+        expect(File).to receive(:read).with(path) { contents }
+        expect(Renderer.contents_of(name)).to eq(contents)
       end
     end
   end
