@@ -4,9 +4,9 @@ require "octopolo/scripts/compare_release"
 module Octopolo
   module Scripts
     describe CompareRelease do
-      let(:config) { stub(:config, github_repo: "tstmedia/ngin") }
-      let(:cli) { stub(:cli) }
-      let(:git) { stub(:git, recent_release_tags: tags) }
+      let(:config) { double(:config, github_repo: "tstmedia/ngin") }
+      let(:cli) { double(:cli) }
+      let(:git) { double(:git, recent_release_tags: tags) }
       let(:tags) { %w(tag1 tag2) }
 
       subject { CompareRelease.new }
@@ -19,16 +19,16 @@ module Octopolo
 
       context "#intialize" do
         it "properly handles no arguments" do
-          subject.start.should be_nil
-          subject.stop.should be_nil
+          expect(subject.start).to be_nil
+          expect(subject.stop).to be_nil
         end
 
         context "with a starting tag only" do
           subject { CompareRelease.new "foo" }
 
           it "sets the start tag only" do
-            subject.start.should == "foo"
-            subject.stop.should be_nil
+            expect(subject.start).to eq("foo")
+            expect(subject.stop).to be_nil
           end
         end
 
@@ -36,17 +36,17 @@ module Octopolo
           subject { CompareRelease.new "foo", "bar" }
 
           it "sets the start and end tags" do
-            subject.start.should == "foo"
-            subject.stop.should == "bar"
+            expect(subject.start).to eq("foo")
+            expect(subject.stop).to eq("bar")
           end
         end
       end
 
       context "#execute" do
         it "prompts for starting and ending tag, then opens the comparei page" do
-          subject.should_receive(:ask_starting_tag)
-          subject.should_receive(:ask_stopping_tag)
-          subject.should_receive(:open_compare_page)
+          expect(subject).to receive(:ask_starting_tag)
+          expect(subject).to receive(:ask_stopping_tag)
+          expect(subject).to receive(:open_compare_page)
 
           subject.execute
         end
@@ -54,14 +54,14 @@ module Octopolo
 
       context "#ask_starting_tag" do
         it "prompts to select from the list of release tags" do
-          cli.should_receive(:ask).with("Start with which tag?", tags) { tags.first }
+          expect(cli).to receive(:ask).with("Start with which tag?", tags) { tags.first }
           subject.ask_starting_tag
-          subject.start.should == tags.first
+          expect(subject.start).to eq(tags.first)
         end
 
         it "does nothing if alread has a starting tag" do
           subject.start = "foo"
-          cli.should_not_receive(:ask)
+          expect(cli).not_to receive(:ask)
           subject.ask_starting_tag
         end
       end
@@ -72,14 +72,14 @@ module Octopolo
         end
 
         it "prompts to select form the list of release tags" do
-          cli.should_receive(:ask).with("Compare from #{subject.start} to which tag?", tags) { tags.last }
+          expect(cli).to receive(:ask).with("Compare from #{subject.start} to which tag?", tags) { tags.last }
           subject.ask_stopping_tag
-          subject.stop.should == tags.last
+          expect(subject.stop).to eq(tags.last)
         end
 
         it "does nothing if already has an ending tag" do
           subject.stop = "bar"
-          cli.should_not_receive(:ask)
+          expect(cli).not_to receive(:ask)
           subject.ask_stopping_tag
         end
       end
@@ -88,9 +88,9 @@ module Octopolo
         let(:url) { "http://example.com/" }
 
         it "copies the compare URL to the clipboard and opens it" do
-          subject.stub(:compare_url) { url }
-          cli.should_receive(:copy_to_clipboard).with(url)
-          cli.should_receive(:open).with(url)
+          allow(subject).to receive(:compare_url) { url }
+          expect(cli).to receive(:copy_to_clipboard).with(url)
+          expect(cli).to receive(:open).with(url)
 
           subject.open_compare_page
         end
@@ -106,7 +106,7 @@ module Octopolo
         end
 
         it "builds the URL from the starting and stopping tags" do
-          subject.compare_url.should == "https://github.com/#{config.github_repo}/compare/#{start}...#{stop}?w=1"
+          expect(subject.compare_url).to eq("https://github.com/#{config.github_repo}/compare/#{start}...#{stop}?w=1")
         end
       end
     end
