@@ -5,29 +5,28 @@ module Octopolo
   module Scripts
     describe PullRequest do
       let(:config) do
-        stub(:config, {
+        double(:config, {
           deploy_branch: "production",
           github_repo: "tstmedia/foo",
-          use_pivotal_tracker: true,
           use_jira: true
         })
       end
-      let(:cli) { stub(:cli) }
+      let(:cli) { double(:cli) }
       let(:current_branch) { "bug-123-something" }
-      let(:git) { stub(:Git, current_branch: current_branch, reserved_branch?: false) }
+      let(:git) { double(:Git, current_branch: current_branch, reserved_branch?: false) }
       let(:pull_request_url) { "http://github.com/tstmedia/octopolo/pull/0" }
-      let(:pull_request) { stub(:pull_request) }
+      let(:pull_request) { double(:pull_request) }
 
       subject { PullRequest.new }
 
       before do
-        PullRequest.any_instance.stub({
+        allow_any_instance_of(PullRequest).to receive_messages({
           :cli => cli,
           :config => config,
           :git => git
         })
 
-        Octopolo::Question.any_instance.stub({
+        allow_any_instance_of(Octopolo::Question).to receive_messages({
           :cli => cli
         })
       end
@@ -44,10 +43,9 @@ module Octopolo
 
       context "#execute" do
         it "if connected to GitHub, asks some questions, creates the pull request, and opens it" do
-          GitHub.should_receive(:connect).and_yield
+          expect(GitHub).to receive(:connect).and_yield
           expect(subject).to receive(:ask_questionnaire)
           expect(subject).to receive(:create_pull_request)
-          expect(subject).to receive(:update_pivotal)
           expect(subject).to receive(:update_jira)
           expect(subject).to receive(:update_labels)
           expect(subject).to receive(:open_in_browser)
@@ -56,7 +54,7 @@ module Octopolo
         end
 
         it "if not connected to GitHub, does nothing" do
-          GitHub.should_receive(:connect) # and not yield, no github credentials
+          expect(GitHub).to receive(:connect) # and not yield, no github credentials
           expect { subject.execute }.to_not raise_error
         end
       end
@@ -65,7 +63,6 @@ module Octopolo
         it "asks appropriate questions to create a pull request" do
           expect(subject).to receive(:announce)
           expect(subject).to receive(:ask_title)
-          expect(subject).to receive(:ask_pivotal_ids)
           expect(subject).to receive(:ask_jira_ids)
           expect(subject).to receive(:ask_labels)
 
@@ -74,24 +71,23 @@ module Octopolo
 
         context "when checking for staging branch" do
           before do
-            subject.stub(:announce)
-            subject.stub(:ask_title)
-            subject.stub(:ask_pivotal_ids)
-            subject.stub(:ask_jira_ids)
-            subject.stub(:ask_labels)
+            allow(subject).to receive(:announce)
+            allow(subject).to receive(:ask_title)
+            allow(subject).to receive(:ask_jira_ids)
+            allow(subject).to receive(:ask_labels)
           end
           it "exits when branch name is reserved" do
-            subject.git.stub(:reserved_branch?).and_return true
+            allow(subject.git).to receive(:reserved_branch?).and_return true
 
-            subject.should_receive(:alert_reserved_and_exit)
+            expect(subject).to receive(:alert_reserved_and_exit)
 
             subject.send(:ask_questionnaire)
           end
 
           it "should not ask if the branch is not staging" do
-            subject.git.stub(:reserved_branch?).and_return false
+            allow(subject.git).to receive(:reserved_branch?).and_return false
 
-            subject.should_not_receive(:alert_reserved_and_exit)
+            expect(subject).not_to receive(:alert_reserved_and_exit)
 
             subject.send(:ask_questionnaire)
           end
@@ -125,16 +121,16 @@ module Octopolo
           let(:current_branch) { 'not_enough'}
 
           it 'does not like other branch format' do
-            subject.git.stub(:reserved_branch?).and_return false
-            cli.should_receive(:say)
+            allow(subject.git).to receive(:reserved_branch?).and_return false
+            expect(cli).to receive(:say)
             expect { subject.send(:infer_questionnaire) }.to raise_error(SystemExit)
           end
         end
 
         it 'does not process reserved' do
-          subject.git.stub(:reserved_branch?).and_return true
-          subject.should_receive(:alert_reserved_and_exit).and_call_original
-          cli.should_receive(:say)
+          allow(subject.git).to receive(:reserved_branch?).and_return true
+          expect(subject).to receive(:alert_reserved_and_exit).and_call_original
+          expect(cli).to receive(:say)
           expect { subject.send(:infer_questionnaire) }.to raise_error(SystemExit)
         end
       end
@@ -145,7 +141,7 @@ module Octopolo
         end
 
         it "displays information about the pull request to be created" do
-          cli.should_receive(:say).with("Preparing a pull request for #{config.github_repo}/#{git.current_branch} to #{config.github_repo}/#{subject.destination_branch}.")
+          expect(cli).to receive(:say).with("Preparing a pull request for #{config.github_repo}/#{git.current_branch} to #{config.github_repo}/#{subject.destination_branch}.")
           subject.send(:announce)
         end
       end
@@ -154,7 +150,7 @@ module Octopolo
         let(:title) { "title" }
 
         it "asks for and captures a title for the pull request" do
-          cli.should_receive(:prompt).with("Title:") { title }
+          expect(cli).to receive(:prompt).with("Title:") { title }
           subject.send(:ask_title)
           expect(subject.title).to eq(title)
         end
@@ -179,38 +175,15 @@ module Octopolo
         end
       end
 
-      context "#ask_pivotal_ids" do
-        let(:ids_with_whitespace) { "123 456" }
-        let(:ids_with_commas) { "234, 567" }
-
-        it "asks for and captures IDs for related pivotal tasks" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_whitespace }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(123 456))
-        end
-
-        it "asks for and captures IDs with commas" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { ids_with_commas }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq(%w(234 567))
-        end
-
-        it "sets to an empty array if not provided an ansswer" do
-          cli.should_receive(:prompt).with("Pivotal Tracker story ID(s):") { "" }
-          subject.send(:ask_pivotal_ids)
-          expect(subject.pivotal_ids).to eq([])
-        end
-      end
-
       context "#create_pull_request" do
-        let(:attributes) { stub(:hash) }
+        let(:attributes) { double(:hash) }
 
         before do
-          subject.stub(:pull_request_attributes) { attributes }
+          allow(subject).to receive(:pull_request_attributes) { attributes }
         end
 
         it "creates and stores the pull request" do
-          GitHub::PullRequest.should_receive(:create).with(config.github_repo, attributes) { pull_request }
+          expect(GitHub::PullRequest).to receive(:create).with(config.github_repo, attributes) { pull_request }
           subject.send(:create_pull_request)
           expect(subject.pull_request).to eq(pull_request)
         end
@@ -223,19 +196,17 @@ module Octopolo
       context "#pull_request_attributes" do
         before do
           subject.title = "title"
-          subject.destination_branch = "some-branch",
-          subject.pivotal_ids = %w(123)
+          subject.destination_branch = "some-branch"
         end
 
         it "combines the anssers with a handful of deault values" do
-          subject.send(:pull_request_attributes).should == {
+          expect(subject.send(:pull_request_attributes)).to eq({
             title: subject.title,
             destination_branch: subject.destination_branch,
             source_branch: git.current_branch,
-            pivotal_ids: subject.pivotal_ids,
             jira_ids: subject.jira_ids,
             editor: nil
-          }
+          })
         end
       end
 
@@ -250,39 +221,10 @@ module Octopolo
         end
       end
 
-      context "#update_pivotal" do
-        before do
-          subject.pivotal_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Pivotal::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Pivotal::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_pivotal)
-        end
-
-      end
-
-      context "#update_jira" do
-        before do
-          subject.jira_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
-        end
-        let(:story_commenter) { stub(perform: true) }
-
-        it "creates a story commenter for each pivotal_id" do
-          Jira::StoryCommenter.should_receive(:new).with("123", "test") { story_commenter }
-          Jira::StoryCommenter.should_receive(:new).with("234", "test") { story_commenter }
-          subject.send(:update_jira)
-        end
-      end
-
       context "#update_labels" do
         before do
           subject.labels = ["high-risk"]
-          subject.pull_request = stub()
+          subject.pull_request = double()
         end
         it "calls update_labels with proper arguments" do
           expect(subject.pull_request).to receive(:add_labels).with(['high-risk'])
@@ -304,12 +246,12 @@ module Octopolo
       context "#open_in_browser" do
         before do
           subject.pull_request = pull_request
-          pull_request.stub(:url) { pull_request_url }
+          allow(pull_request).to receive(:url) { pull_request_url }
         end
 
         it "copies the pull request's URL to the clipboard and opens it in the browser" do
-          cli.should_receive(:copy_to_clipboard) { pull_request.url}
-          cli.should_receive(:open) { pull_request.url }
+          expect(cli).to receive(:copy_to_clipboard) { pull_request.url}
+          expect(cli).to receive(:open) { pull_request.url }
           subject.send(:open_in_browser)
         end
       end
