@@ -4,15 +4,15 @@ require "octopolo/scripts/stage_up"
 module Octopolo
   module Scripts
     describe StageUp do
-      let(:cli) { stub(:Cli) }
-      before { StageUp.any_instance.stub(:cli => cli) }
+      let(:cli) { double(:Cli) }
+      before { allow_any_instance_of(StageUp).to receive_messages(:cli => cli) }
 
       context "#execute" do
         context "with a PR passed in via the command args" do
           subject { StageUp.new 42 }
 
           it "delegates the work to PullRequestMerger" do
-            PullRequestMerger.should_receive(:perform).with(Git::STAGING_PREFIX, 42)
+            expect(PullRequestMerger).to receive(:perform).with(Git::STAGING_PREFIX, 42)
             subject.execute
           end
         end
@@ -22,36 +22,36 @@ module Octopolo
 
           context "with an existing PR for the current branch" do
             before do
-              GitHub::PullRequest.should_receive(:current) { GitHub::PullRequest.new("account/repo", 7) }
+              expect(GitHub::PullRequest).to receive(:current) { GitHub::PullRequest.new("account/repo", 7) }
             end
 
             it "takes the pull requests ID from the current branch" do
-              PullRequestMerger.should_receive(:perform).with(Git::STAGING_PREFIX, 7)
+              expect(PullRequestMerger).to receive(:perform).with(Git::STAGING_PREFIX, 7)
               subject.execute
             end
           end
 
           context "without an existing PR for the current branch" do
             before do
-              GitHub::PullRequest.should_receive(:current) { nil }
+              expect(GitHub::PullRequest).to receive(:current) { nil }
             end
 
             context "with a PR passed in through the cli" do
               before do
-                cli.should_receive(:prompt)
+                expect(cli).to receive(:prompt)
                    .with("Pull Request ID: ")
                    .and_return("42")
               end
 
               it "delegates the work to PullRequestMerger" do
-                PullRequestMerger.should_receive(:perform).with(Git::STAGING_PREFIX, 42)
+                expect(PullRequestMerger).to receive(:perform).with(Git::STAGING_PREFIX, 42)
                 subject.execute
               end
             end
 
             context "with no PR passed in from the cli" do
               before do
-                cli.should_receive(:prompt)
+                expect(cli).to receive(:prompt)
                    .with("Pull Request ID: ")
                    .and_return("foo")
               end
