@@ -5,14 +5,14 @@ module Octopolo
   module Scripts
     describe StaleBranches do
       # stub any attributes in the config that you need
-      let(:config) { stub(:config, :deploy_branch => "production", :branches_to_keep => %w(master production staging)) }
-      let(:cli) { stub(:cli) }
-      let(:git) { stub(:Git)}
+      let(:config) { double(:config, :deploy_branch => "production", :branches_to_keep => %w(master production staging)) }
+      let(:cli) { double(:cli) }
+      let(:git) { double(:Git)}
 
       subject { StaleBranches.new }
 
       before do
-        StaleBranches.any_instance.stub({
+        allow_any_instance_of(StaleBranches).to receive_messages({
           :cli => cli,
           :config => config,
           :git => git
@@ -21,57 +21,57 @@ module Octopolo
 
       context "#new" do
         it "accepts a delete attribute to trigger deletes" do
-          expect(StaleBranches.new(true).delete?).to be_true
+          expect(StaleBranches.new(true).delete?).to be_truthy
         end
 
         it "defaults to not delete" do
-          expect(subject.delete?).to be_false
+          expect(subject.delete?).to be_falsey
         end
       end
 
       context "#execute" do
         it "displays the stale branches if not deleting" do
           subject.delete = false
-          subject.should_receive(:display_stale_branches)
-          subject.should_not_receive(:delete_stale_branches)
+          expect(subject).to receive(:display_stale_branches)
+          expect(subject).not_to receive(:delete_stale_branches)
           subject.execute
         end
 
         it "deletes the branches if deleting" do
           subject.delete = true
-          subject.should_not_receive(:display_stale_branches)
-          subject.should_receive(:delete_stale_branches)
+          expect(subject).not_to receive(:display_stale_branches)
+          expect(subject).to receive(:delete_stale_branches)
           subject.execute
         end
       end
 
       context "#display_stale_branches" do
         before do
-          subject.stub(:stale_branches) { %w(foo bar) }
+          allow(subject).to receive(:stale_branches) { %w(foo bar) }
         end
 
         it "displays a list of stale branches" do
-          cli.should_receive(:say).with("* foo")
-          cli.should_receive(:say).with("* bar")
+          expect(cli).to receive(:say).with("* foo")
+          expect(cli).to receive(:say).with("* bar")
           subject.send(:display_stale_branches)
         end
       end
 
       context "#delete_stale_branches" do
         before do
-          subject.stub(:stale_branches) { %w(foo bar) }
+          allow(subject).to receive(:stale_branches) { %w(foo bar) }
         end
 
         it "deletes each stale branch" do
-          git.should_receive(:delete_branch).with("foo")
-          git.should_receive(:delete_branch).with("bar")
+          expect(git).to receive(:delete_branch).with("foo")
+          expect(git).to receive(:delete_branch).with("bar")
           subject.send(:delete_stale_branches)
         end
       end
 
       context "#stale_branches" do
         it "fetches from the git wrapper" do
-          git.should_receive(:stale_branches).with(config.deploy_branch, config.branches_to_keep)
+          expect(git).to receive(:stale_branches).with(config.deploy_branch, config.branches_to_keep)
           subject.send(:stale_branches)
         end
       end
