@@ -2,6 +2,6 @@ require "spec_helper"
 
 describe Octopolo do
   it "has a VERSION" do
-    Octopolo::VERSION.should_not be_nil
+    expect(Octopolo::VERSION).not_to be_nil
   end
 end
