@@ -1,3 +1,3 @@
 module Octopolo
-  VERSION = "1.11.1"
+  VERSION = "2.0.0"
 end
